+++ packages/wonder-blocks-banner/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @khanacademy/wonder-blocks-banner
 
+## 3.1.14
+
+### Patch Changes
+
+-   Updated dependencies [c1110599]
+    -   @khanacademy/wonder-blocks-icon@4.2.0
+    -   @khanacademy/wonder-blocks-button@6.3.12
+    -   @khanacademy/wonder-blocks-icon-button@5.6.2
+    -   @khanacademy/wonder-blocks-link@6.1.10
+
+## 3.1.13
+
+### Patch Changes
+
+-   Updated dependencies [6999fd39]
+    -   @khanacademy/wonder-blocks-tokens@2.1.0
+    -   @khanacademy/wonder-blocks-button@6.3.11
+    -   @khanacademy/wonder-blocks-icon-button@5.6.1
+    -   @khanacademy/wonder-blocks-link@6.1.9
+
 ## 3.1.12
 
 ### Patch Changes