+++ packages/wonder-blocks-dropdown/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @khanacademy/wonder-blocks-dropdown
 
+## 5.8.0
+
+### Minor Changes
+
+-   2b8424ca: Allow use of JSX Element as label in SingleSelect and MultiSelect
+
+## 5.7.0
+
+### Minor Changes
+
+-   c512e76e: Add `startIcon` prop to Combobox
+
+### Patch Changes
+
+-   Updated dependencies [b5cad0b1]
+    -   @khanacademy/wonder-blocks-modal@5.1.17
+
+## 5.6.3
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-search-field@2.3.6
+
+## 5.6.2
+
+### Patch Changes
+
+-   Updated dependencies [c1110599]
+    -   @khanacademy/wonder-blocks-icon@4.2.0
+    -   @khanacademy/wonder-blocks-search-field@2.3.5
+    -   @khanacademy/wonder-blocks-modal@5.1.16
+    -   @khanacademy/wonder-blocks-pill@2.5.3
+
+## 5.6.1
+
+### Patch Changes
+
+-   Updated dependencies [6999fd39]
+    -   @khanacademy/wonder-blocks-tokens@2.1.0
+    -   @khanacademy/wonder-blocks-cell@3.4.6
+    -   @khanacademy/wonder-blocks-clickable@4.2.9
+    -   @khanacademy/wonder-blocks-layout@2.2.2
+    -   @khanacademy/wonder-blocks-modal@5.1.15
+    -   @khanacademy/wonder-blocks-pill@2.5.2
+    -   @khanacademy/wonder-blocks-search-field@2.3.4
+
 ## 5.6.0
 
 ### Minor Changes