+++ packages/wonder-blocks-modal/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @khanacademy/wonder-blocks-modal
 
+## 5.1.17
+
+### Patch Changes
+
+-   b5cad0b1: Use h2 tag in modal header instead of h3
+
+## 5.1.16
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-icon-button@5.6.2
+
+## 5.1.15
+
+### Patch Changes
+
+-   Updated dependencies [6999fd39]
+    -   @khanacademy/wonder-blocks-tokens@2.1.0
+    -   @khanacademy/wonder-blocks-breadcrumbs@2.2.8
+    -   @khanacademy/wonder-blocks-icon-button@5.6.1
+    -   @khanacademy/wonder-blocks-layout@2.2.2
+
 ## 5.1.14
 
 ### Patch Changes